refactor: refactored directory for MainLayout
