+++ frontend/src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Routes, Route } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 import { GlobalStyles, theme } from "./styles/globalStyles";
-import MainLayout from "./main/Layouts/MainLayout";
+import MainLayout from "./main/Layouts/Main/MainLayout";
 
 import Contact from "./main/pages/Contact";
 import { AdminAuthLayout } from "./main/components";
