+++ frontend/src/main/Layouts/Main/Partials/Header.jsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import styled from "styled-components";
 import { useLocation } from "react-router-dom";
 
-import logo from "../../../assets/images/logo.svg";
-import menuIcon from "../../../assets/icons/menu.svg";
-import { Container, Button } from "../../../styles/reusableElements.styled";
+import logo from "../../../../assets/images/logo.svg";
+import menuIcon from "../../../../assets/icons/menu.svg";
+import { Container, Button } from "../../../../styles/reusableElements.styled";
 import NavBar from "./NavBar";
 import { Link } from "react-router-dom";
 import MobileNav from "./MobileNav";
