+++ frontend/src/main/Layouts/Main/Partials/NavBar.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
 
-import arrowDown from "../../../assets/icons/arrow-down.svg";
+import arrowDown from "../../../../assets/icons/arrow-down.svg";
 
 const NavBar = () => {
   return (
